refactor activity flow page and fixed bug on activity indicator
