+++ lib/activity_flow/activity_flow_widget.dart
@@ -1,9 +1,7 @@
 import 'package:flutter/material.dart';
 
-import 'package:festival_flutterturkiye_org/activity_flow/activity_flow_icon.dart';
-import 'package:festival_flutterturkiye_org/activity_flow/activity_flow_line_indicator.dart';
-import 'package:festival_flutterturkiye_org/activity_flow/activity_flow_subtitle.dart';
-import 'package:festival_flutterturkiye_org/activity_flow/activity_flow_title.dart';
+import 'package:festival_flutterturkiye_org/core/utils/image_assets.dart';
+import 'package:festival_flutterturkiye_org/core/utils/theme_helper.dart';
 
 class ActivityFlowWidget extends StatelessWidget {
   const ActivityFlowWidget({
@@ -40,9 +38,9 @@ class ActivityFlowWidget extends StatelessWidget {
                 isLastIndex: isLastIndex,
                 iconPath: iconPath,
               ),
-              const SizedBox(height: 10),
-              ActivityFlowTitle(title: title),
-              ActivityFlowSubtitle(subtitle: subtitle),
+              const SizedBox(height: 12),
+              _ActivityFlowTitle(title: title),
+              _ActivityFlowTitle(title: subtitle, isTitle: false),
             ],
           ),
         ),
@@ -69,12 +67,105 @@ class _BuildRowHeader extends StatelessWidget {
         crossAxisAlignment:
             (index.isEven) ? CrossAxisAlignment.start : CrossAxisAlignment.end,
         children: [
-          ActivityFlowIcon(iconPath: iconPath),
-          const SizedBox(width: 10),
-          ActivityFlowLineIndicatorBuilder(
+          _ActivityFlowIcon(iconPath: iconPath),
+          const SizedBox(width: 12),
+          _ActivityFlowLineIndicatorBuilder(
             index: index,
             isLastIndex: isLastIndex,
           ),
         ],
+      );
+}
+
+class _ActivityFlowTitle extends StatelessWidget {
+  const _ActivityFlowTitle({
+    @required this.title,
+    this.isTitle = true,
+    Key key,
+  })  : assert(title != null),
+        assert(isTitle != null),
+        super(key: key);
+
+  final String title;
+  final bool isTitle;
+
+  @override
+  Widget build(BuildContext context) => Text(
+        title,
+        textAlign: TextAlign.start,
+        style: isTitle
+            ? const TextStyle(
+                color: ThemeHelper.activityFlowColor,
+                fontWeight: FontWeight.bold,
+                fontSize: 18,
+              )
+            : const TextStyle(
+                color: ThemeHelper.activityFlowColor,
+                fontWeight: FontWeight.w200,
+                fontSize: 16,
+              ),
+      );
+}
+
+class _ActivityFlowIcon extends StatelessWidget {
+  const _ActivityFlowIcon({
+    @required this.iconPath,
+    Key key,
+  })  : assert(iconPath != null),
+        super(key: key);
+
+  final String iconPath;
+
+  @override
+  Widget build(BuildContext context) => CircleAvatar(
+        radius: 25,
+        backgroundColor: const Color(0xFFF4F5F7),
+        child: Image.asset(iconPath, scale: 1.5),
+      );
+}
+
+class _ActivityFlowLineIndicatorBuilder extends StatelessWidget {
+  const _ActivityFlowLineIndicatorBuilder({
+    @required this.index,
+    @required this.isLastIndex,
+    Key key,
+  })  : assert(index != null),
+        assert(isLastIndex != null),
+        super(key: key);
+
+  final int index;
+  final bool isLastIndex;
+
+  @override
+  Widget build(BuildContext context) {
+    if (isLastIndex) {
+      print('isLastIndex  -> $index');
+      return const SizedBox.shrink();
+    } else if (index.isEven) {
+      print('isEven  -> $index');
+      return const _ActivityFlowLineIndicator(ImageAssets.lineIndicatorTop);
+    } else if (index.isOdd) {
+      print('isOdd  -> $index');
+      return const _ActivityFlowLineIndicator(ImageAssets.lineIndicatorBottom);
+    } else {
+      print('else  -> $index');
+      return const SizedBox.shrink();
+    }
+  }
+}
+
+class _ActivityFlowLineIndicator extends StatelessWidget {
+  const _ActivityFlowLineIndicator(
+    this.assetsPath, {
+    Key key,
+  })  : assert(assetsPath != null),
+        super(key: key);
+
+  final String assetsPath;
+
+  @override
+  Widget build(BuildContext context) => Padding(
+        padding: const EdgeInsets.all(12),
+        child: Image.asset(assetsPath, scale: 1.1),
       );
 }
